Bug 1840061 - Adds PDF viewer - Open in app button UI tests
